+++ packages/atproto_core/lib/src/client/client_context.dart
@@ -11,6 +11,8 @@ import 'package:xrpc/xrpc.dart' as xrpc;
 
 // 🌎 Project imports:
 import '../config/retry_config.dart';
+import '../pagination/pageable.dart';
+import '../pagination/pagination.dart';
 import 'challenge.dart';
 import 'client_resolver.dart';
 import 'retry_policy.dart';
@@ -35,6 +37,17 @@ abstract class ClientContext {
     final xrpc.Protocol? protocol,
     required final String service,
     final Map<String, dynamic>? parameters,
+    final xrpc.To<T>? to,
+    final xrpc.ResponseAdaptor? adaptor,
+    final xrpc.GetClient? getClient,
+  });
+
+  Pagination paginate<T extends Pageable>(
+    final xrpc.NSID methodId, {
+    required UserContext userContext,
+    final xrpc.Protocol? protocol,
+    required final String service,
+    required final Map<String, dynamic> parameters,
     final xrpc.To<T>? to,
     final xrpc.ResponseAdaptor? adaptor,
     final xrpc.GetClient? getClient,
@@ -106,6 +119,31 @@ class _ClientContext implements ClientContext {
       await _challenge.execute(
         _clientResolver.execute(userContext),
         (client) async => await client.get(
+          methodId,
+          protocol: protocol,
+          service: service,
+          parameters: parameters,
+          to: to,
+          adaptor: adaptor,
+          timeout: timeout,
+          getClient: getClient,
+        ),
+      );
+
+  @override
+  Pagination paginate<T extends Pageable>(
+    final xrpc.NSID methodId, {
+    required UserContext userContext,
+    final xrpc.Protocol? protocol,
+    required final String service,
+    required final Map<String, dynamic> parameters,
+    final xrpc.To<T>? to,
+    final xrpc.ResponseAdaptor? adaptor,
+    final xrpc.GetClient? getClient,
+  }) =>
+      _challenge.execute(
+        _clientResolver.execute(userContext),
+        (client) => client.paginate(
           methodId,
           protocol: protocol,
           service: service,